Adjusted moon station position
Moved the moon station so that it interacts with the player casting animation in a better way.
Sorry if it comes across that I keep committing on tiny little tweaks unnecessarily, I am just aware that other developers on the team may feel it necessary to have the most up to date version of something I have implemented, even if all I have done is move a sprite a few measures to the left.